auto install solc versions in detector tests
